components are life
