+++ README.md
@@ -1,52 +1,7 @@
 urg_node
 ===================
 
-ROS 2 wrapper for the Hokuyo urg_c library.
-
-## Build instructions
-Install ROS 2 as per instructions [here](https://index.ros.org/doc/ros2/Installation/#installationguide)
-
-Create a folder for urg and clone this repo
-
-```
-mkdir -p ~/ros2_ws/src/urg
-cd ~/ros2_ws/src/urg
-git clone -b ros2-devel https://github.com/hopetechnik/urg_node.git
-```
-Clone dependencies, you might need to install the vcs command to do so :
-```
-sudo apt install python3-vcstool
-```
-then clone :
-```
-cd ~/ros2_ws/src/urg/urg_node
-vcs import ../ < urg.repos
-```
-
-Finally build :
-To build everything :
-```
-cd ~/ros2_ws
-colcon build --symlink-install
-```
-
-To build urg_node only :
-```
-cd ~/ros2_ws
-colcon build --packages-select urg_node
-```
-
-To build urg_node and all its dependencies : 
-```
-cd ~/ros2_ws
-colcon build --packages-up-to urg_node
-```
-
-More information about colcon is available [here](https://media.readthedocs.org/pdf/colcon/latest/colcon.pdf)
-
-
 #### LaserScan Visualization
-
 
 Until the launch API is sorted out, there are two ways to view the laserscan in RViz:
 
@@ -64,7 +19,7 @@ where "laser" is the frame id, which you can set in your urg_node yaml file
 ros2 run robot_state_publisher robot_state_publisher <path to urdf file>
 ```
 
-A urdf file is already included and if you have succesfully ran an `colcon build`, the path to urdf file will be :
+A urdf file is already included and if you have succesfully ran an `colcon build`, the path to urdf file will be:
 
 ```
 <path to ros2_ws>/install/share/urg_node/launch/hokuyo_laser.urdf
@@ -75,6 +30,7 @@ A urdf file is already included and if you have succesfully ran an `colcon build
 
 A YAML file example is included in the launch folder, all the available parameters are listed in it.
 For example (note that the serial_port is commented because you can't set a param with an empty string):
+
 ```
 urg_node:
   ros__parameters:
@@ -98,12 +54,14 @@ urg_node:
 ```
 
 To give parameters to urg_node :
+
 ```
 ros2 run urg_node urg_node __params:=path/to/my/file.yaml
 ```
 
 You can reconfigure parameters while the node is launched.
 For now, you can only reconfigure the following parameters:
+
 ```
 laser_frame_id
 error_limit
@@ -113,6 +71,7 @@ angle_min
 cluster
 skip
 ```
+
 For example to reconfigure the cluster parameter using command line :
 ```
 ros2 param set /urg_node cluster 1
@@ -121,41 +80,49 @@ ros2 param set /urg_node cluster 1
 #### How to use the ust-20lx (and other ethernet based laser)
 
 To use ust-20lx, you need to be on the same subnet as the laser.
-The ust-20lx default ip is 192.168.0.10, so you might need to change your ip, for something on the same subnet. 
+The ust-20lx default ip is 192.168.0.10, so you might need to change your ip, for something on the same subnet.
 
 On Ubuntu :
-- go to the network settings 
-- settings of the wired connection 
-- under IPv4, change the IPv4 method from automatic to Manual and under Addresses, set the values :
+- go to the network settings
+- settings of the wired connection
+- under IPv4, change the IPv4 method from automatic to Manual and under Addresses, set the values:
+
 ```
     Address : 192.168.0.15
     Netmask : 255.255.255.0
     Gateway : 192.168.0.1
 ```
+
 - Turn off and on the connection or wait a bit for the change to apply
 - To check if it worked, open a terminal and type
+
 ```
 ifconfig
 ```
 
 - Under eth0 (or maybe something like enpxxxxx), your IP should be 192.168.0.15.
 You should now be able to ping the ust-20lx at its address (by default 192.168.0.10)
+
 ```
 ping 192.168.0.10
 ```
+
 If you don't receive any answer, you might have a connection problem or the IP of your laser might have been change, either find it and go on the same subnet or reset it.
 
-- Once you can ping the laser, you can launch the urg_node :
+- Once you can ping the laser, you can launch the urg_node:
+
 ```
 ros2 run urg_node urg_node __params:=path/to/my/file.yaml
 ```
 
-then the static publisher :
+then the static publisher:
+
 ```
 ros2 run tf2_ros static_transform_publisher 0 0 0 0 0 0 "world" "laser"
 ```
 
-and rviz :
+and rviz:
+
 ```
 ros2 run rviz2 rviz2
 ```
