+++ launch/urg_node_ethernet.yaml
@@ -4,7 +4,6 @@ urg_node:
     angle_min: -3.14
     ip_address: "192.168.0.10"
     ip_port: 10940
-    #serial_port: ""
     serial_baud: 115200
     laser_frame_id: "laser"
     calibrate_time: false
