--- urg.repos
@@ -1,25 +1,0 @@
-repositories:
-  diagnostics:
-    type: git
-    url: https://github.com/MarcTestier/diagnostics
-    version: ros2-devel
-  laser_proc:
-    type: git
-    url: https://github.com/MarcTestier/laser_proc
-    version: ros2-devel
-  ros2_console:
-    type: git
-    url: https://github.com/MarcTestier/ros2_console
-    version: master
-  ros2_time:
-    type: git
-    url: https://github.com/MarcTestier/ros2_time
-    version: master
-  urg_c:
-    type: git
-    url: https://github.com/MarcTestier/urg_c
-    version: ros2-devel
-  urg_node_msgs:
-    type: git
-    url: https://github.com/MarcTestier/urg_node_msgs
-    version: master
